Use reflection to get DB models
While we transition to the new servers I will be making many migrations to the
database, things such as changing table types, making names of columns and
tables consistent, adding tables and columns, etc. In order to do this I'm using
a migration tool which is separate from the pipeline as the pipeline does not do
a good job with these sorts of changes. To prevent myself from having to worry
about keeping the model definitions up-to-date along with the code to use them,
I'm using reflection to generate the models. This way the truth is always the
database structure and the models will always match the database. This does not
solve the problem of not having the correct code to use the models, but nothing
would.

In addition, I've made some other changes to getting the engine and session. We
shouldn't use a single session for the entire pipeline so I've included the
session maker object in the top level. In addition, we use logging not print
statements to show what we are connecting to. And finally, we use the given DB
uri and not generate one for us. I am ignoring the environment data however.
This may be a bad idea.