Add some helpers
Accessing the database and the like is a very common activity so I'm adding
helpers for that stuff. The database help allows one to do things like:

    with self.session() as session:
        code

making it easier to have a session per step and the like. It will handle rolling
back on exceptions and has some generic logging for such things. It also has
some simple code for storing data.

I've also added a cif file finder helper. This shows up in a couple places so
I've added code for it. Simple code, nothing special but useful. There are tests
for it as well.