Add first pass at computing nt-nt correspondences
I've added a module that contains logic for computing the nt-nt correspondences.
The basic idea is that we get a PDB, get the representative pdb then align the
sequences of the two structures. We use the longest chain of each for the
alignment. The actual alignment is done in a java program. We access it by
sending a web request. The python code handles figuring out what to align sends
the data and then stores the alignment. This is simply a first pass at the
structure that has some tests to show basic parts seem to work.

Also, there is some sql written as a string. This may not be ideal, but the
reason I'm using it is because sqlalchemy refuses to reflect in views. This is
annoying but injection and the like isn't a risk here as this isn't user facing.
Plus it does escape and such for us. A bit annoying but not the end of the
world.