Refactor and test nt_nt_correspondence
I've refactored a lot of code in this module as well as added some tests for it.
I think the code is a bit more straightforward and reliable now. I still need
tests for making the request and dealing with failed requests, however, that
requires some mocking which I haven't quite figured out yet.