Add retry helper and ftp helper
The retry helper is something to help retry requests that can fail. For example,
making web requests and fetching something from the FTP site. These things can
fail sometimes so we need to allow for retries of them. Also, I've added an FTP
helper to make it easier to fetch files from an FTP site.