Add initial outline of an nt quality importer
I've added the initial pass at outlining something that will import NT quality
data. The idea is to fetch the file from the FTP site, unzip it, then parse the
XML to create nt level annotations. I haven't examined files to see all that
needs to be added, but this is the initial pass at the structure that will be
used.