Add initial pass at LoopLoopCorresopndence
This loader will compute the correspondences between loops of two structures. We
will also compute discrepancy for loops that are exact. This is then stored in
the database. Much of this code comes from some previous work I did to make a
quick and dirty test of this method. This needs many more tests.