Attempt to refactor downloading to use webhelper
The webhelper has retries which is cleaner than the current attempt. Also, I
have run into errors that occur outside of the current try block for getting
files. This refactor moves code around we the whole thing is in a try block and
we get retries when trying to get the files.