Add a core.py
Core.py is meant to be a replacement for MotifAtlasBaseClass.py. That is to say
it will contain the classes that all loaders (and exporters eventually) will
inherit. It is meant to centralize logic about whether or not we should compute
some annotations. It also contains things for handling sessions, storing to the
database and matlab interfaces. My hope is that by doing this it is much simpler
and more reliable to add new importers. Right now any class using it need only
add 3 methods and will get lots of useful functionality. Doing this will provide
a uniform interface to all loaders, which will be very useful. It differs from
the current base class in that it is a new style class and does not contain all
functionality.

It will eventually supplant MotifAtlasBaseClass but for now it cannot. Because
it is a new style class the initialization of child classes will be different.
This will not contain all functionality of the current base class. For example,
it will not setup logging, this will be done at the application level. However,
we need to add things such as emailing in the future. Once this module is more
mature I will begin to move older classes over to it. Probably starting with
simple ones like distances and coordinates. For now I will use it on the new
parts of the pipeline I'm developing, such as storing polymer alignments. These
parts aren't ready to be in the core but are useful and will provide a good test
case for this module.