Overly large improvement commit
I've done several things in this commit (sorry). First I've worked on getting
correspondence.info and correspondence.nts working. They seem to be correct now.
Also, I've added a bin/pipeline script. The idea behind this script is to
provide an entry point for the new style classes that inherit from core.py. We
also get to skip out on adding boiler plate to every module we want to run and
get a consistent method of running stuff that is documented and uses args and
options.

I have also reorganized the utils stuff. This stuff has been useful but spread
out and causes issues when importing things. Notably doing import utils has
different effects if you are in a subpackage like 'correspondence' vs in
pymotifs itself. This is due to pymotifs not being a python module as it has no
__init__.py file. I could add one to fix this but that would require changing
the importing for every other class.

I've also fixed problems in running PdbDownloader. Doing this meant adding a
Gzip helper so we can make web requests for gzip stuff and get the unzipped
content back. This prompted a change in how the parser in webrequest helper is
called. I've changed it to now use the entire response instead of just the text.
This is more flexible but does break some previous things.

I have also commented out parts of core.py temporarily. Notably, the part that
marks a stage as done. I haven't updated the database yet and am not 100% on the
design for the new parts, so until then I will keep this commented.