Only map things from 1_555 and model 1
This is limited but consistent with exp_seq/positions. Future work will have to
expand this code to something better but this does allow a start and will work
for many interesting structures. If we did not do this then we go errors when
trying to work with a chain that does not exist.