Deal with chains that have no valid mapping
Some chains do not have a valid mapping. This does happen sometimes but they
seem to be uninteresting chains, things like antibiotics and the like. I think.
This commit adds error handling for such things. However the errors thrown are
very generic so this should be fixed and the handling fixed as well.