Improvements to core
It is better to define a logger property for core. This means we can set it's
name to be the name of the stage we are running. This way the logging messages
are clear as to what is happening. Though this does obscure the source of the
error, which may prove to be an issue for people unfamiliar with the structure.
However, it does make it much easier for someone who is familiar to the
structure to know what stage is running and what has failed.

Also, we change transform to provide a list. This lets me improve the logging
messages to include things like how far along the process we are. This is
very useful when watching the pipeline.

Finally, I add skip exceptions. These allow a stage to skip parts when needed.
This is meant to be used when we have run into an error condition that prevents
the stage from continuing with this specific input but we still want to keep
going. For example we may require nt level data on a stage, but we haven't
imported it for a particular PDB. In this case we should no longer process the
PDB but keep trying with the rest of the PDBs. There are two types of skip
exceptions, SkipPdb which skips an entire PDB and SkipValue which skips only a
single transformed value for a PDB.