Improve reflected classes
We use reflection to get all ORM classes and this mostly works, with a few
exceptions. Those exceptions are classes which have methods that the pipeline
depends on. To deal with this I've first added those classes back in. However,
I've fixed their names to match the standard naming scheme. I have also changed
the reflect function so that it will not attempt to reflect these tables. I have
also added a metadata as this is needed to define the function that determines
if we should reflect or not. Hopefully, this will allow the nr and motif atlas
to function.