Begin to outline code to group chains
The main work here is to create a class which can be used to group all RNA
chains into non-redundant groups. I'm not calling these classes because NR
classes have ids, are attached to releases and have differences computed between
them and other classes. This code will do none of that. It is meant to take a
list of pdbs and produce the grouping for all chains.

Right now this code is just an outline of what needs to be done and roughly how
things will work. Many methods are just stubs and none are tested. This will be
the starting point for this work. The design is based off the notes Dr. Zirbel
gave me in an email "Re: The nr pipeline" (id:548EFDFD.60406@bgsu.edu).