Outline the initial stage to get species mapping
We need something to take the given organism information for a chain and
determine the species it comes from. This is that stage. The current idea is
that it will use the taxonomyId field to query NCBI for taxon information and
store it in the database. There is some complication in making sure we don't get
banned from NCBI for too many requests and such.