Add tests for finding autonomous groups of chains
We can now detect simple groups of autonomous chains. This commit adds code and
tests for finding this with two easy examples, one is 2AW7 which only has one
RNA chain and the other is 1ET4 which Dr. Zirbel has suggested as a good example
to work with. 1ET4 has several rna chains which fall into 3 groups. We can find
this easily enough.