Add tests for nr chains
We now get the nr chains for a series of structures. I've also added test
specific to selecting the best chain for a series of structures. The method is
pretty basic currently, but it does seem to work on these examples. One current
issue is we do not yet deal with source information. Should we group chains with
different sources?. We will need to decide this and other issues, but we require
examples. Some LSU's are good examples as these often have a tRNA from a
different organism.