Add test and fix bug with sorting entities
It is possible that if we do not sort entities that we will group chains
incorrectly. For example if A is entity 1 and B is 2, C is 2 and D is 1 and the
groups are A,B and C,D then the unsorted entities will be 1,2 and 2,1 which will
be treated differently when grouping. We will then get two groups of the same
thing instead of just one group with two examples. Instead of fixing this while
grouping I simply change the merging of chain data to always sort the entities.
This forces the entities into a nice reliable order that will prevent such
issues.