Add some failing tests for loading source
The source is key in the next step in the processing. I don't yet have
determining worked out, however these are some good tests to show what the
correct behavior should be. We add tests for things that have several sources,
synthetic sources, and sources where we need to go from a sub-species to a
species level.