Make UnitInfo work again
I had a bunch of incomplete work on caching that I'm simply not using now. This
will slow things down but a slow pipeline is better than one that does not work.
I have made sure that info works well enough with this. I'm going to work
through the other important basic parts and get them working as well.