Finally cleaned up everything
This is a cleaner version where we can run nosetests now. We don't have failures
but we do have lots of expected skips to mark things to do. The coverage is
really poor but this is a good place to start the work on fixing that. I've also
fixed a bug in getting pdb info.