Add some simple tests for determining species
This adds some very simple and preliminary tests for getting the species of
things. They should be improved later but this appears to be enough to get the
species mapping stage working for simple cases. We use a trimmed can example of
data from NCBI to test some parsing logic. We look to see if we can use the
species given as well as look up the taxonomy tree for the correct species. I
don't yet test cases of unknown ids or things above the species level. Finally,
we need to test that we do get the correct missing ids.