Add a dry run option to storing
We add an extra keyword argument to store which causes the method to not add
anything to the database, but instead produce a debug level log about what was
going to be stored.

This option will be very useful for exploratory debugging. We can run stages on
large inputs without actually changing the db (which can be messy to clean up
sometimes) and see if everything gets processed. This will be a good way to find
examples that currently break things.