Fix issue with loading given multiple classes
If there was more than one class in the module it would load the first one,
instead of trying to find which one was the stage. That was not good behavior,
so I've fixed it. In addition, I've changed MultiLoader to be a subclass of
stage to allow it to load as well.