Make retry helper only log on last fail by default
It makes the logs really noisy in some cases. For example, when getting quality
for something that doesn't exist, we get a lot of exception logs. This way we
only get one on the last failure by default. We can still configure things so
that they log every time using an initialization parameter.