Add a loader for obsolete info
This can load up information about obsolete data and merge it into our database.
This required a bunch of changes to core, notably adding the ability to merge.
I've added this as well as a bunch of tests for storing things in the database.
I've also removed the ability to deal with nested generators as that was no
needed and too complex.