Use .one() in tests
This is better because .first() will work if there is one or more, while .one()
assumes there is only one. Thus our tests are more sensitive to what is going on
in the DB and should prevent us from passing if the DB is not clean. However, it
will cause extra failures if the DB is messy, but this is ok I think.