Make SpeciesMapping a MassLoader
This should be a mass loader as it doesn't really need input from PDB files to
run. In addition, I've made it merge_data, as this is generally what a mass
loader should do. Plus it filters to make sure that there is data available, before
adding to the list of data to return. Finally I've fixed the table name for the
species mapping table.