Create a working chain info loader
This loader can get information about chains from PDB using their web interface.
I've added tests for it and realized that we need to not just hetId in the
query. This is because if we use hetId we get one line per het type, which we do
not want. This leads to more than one line per chain in the case of 1S72. My
tests picked up on this and I don't think it's a good idea. Plus we don't use
het_id in the database for anything, and if we do it's wrong. Thus I'm not
including it in the update pipeline.