An initial pass at creating a loop loader
This is my first pass at creating something which will extract loops from PDB
files. This is just a modification of the previous LoopExtractor module to use
the newer setup. It uses matlab to get the loops, then processes them a bit to
assign filenames and loop ids, uses matlab to save the loop files to the correct
location and then saves the loop information in the database.