Attempt to generate loop ids correctly
We now attempt to generate loop ids correctly. Before we had an issue where I
followed the previous logic that relies upon saving each loop constantly so we
can use a database query to get the count. As we no longer save until the end
this does not work. I've fixed things to generate a mapping which can be used to
get old ids and the pass along a count to get a new one.