Rework loop extractor
We must send strings, not unicode strings to matlab. If we do we get an
exception about missing floats for some reason. This is now fixed by casting to
str before sending it to matlab.

We also had to reorganize the loader so that saving happens in the same matlab
session as the loading, we can't pass things between sessions (I think). This
means we got rid of the separate save function and just made it part of getting
loops. That also promoted reworking how ids are generated as well.

In addition, I've fixed some tests for the loader.