Deal with issue with deleting using joins
sqlalchemy does not delete from MySQL using joins. Odd, but we can work around
that by using 2 queries to delete. This does so and adds a smoke test for this.
In addition, we add a skip test that will be implemented once we have a test
database for real deleting.