Use loop_name to get ids
We must use loop name because nt_ids are being moved from old style to unit ids.
This means that the ids will not be usable to find existing loops. As we don't
want to have to wipe everything this is a better choice. We can update
everything this way though.