Add a utility class for manipulating releases
It is very useful to get current releases for various things, like the nr, motif
and loops. We also need to get the next release in various ways. This can check
do such things. I've included some simple tests for this stuff. It does not yet
deal with badly formatted ids or for the first time this is run on an empty
database.