Rework transform methods
When we transform it is better to not parse the cif file right away. If we do
so, we run the problem of spending a long time parsing something that isn't
needed. Thus we no longer transform into a fr3d Structure. However, we want to
be able to parse such things to data methods and the like for testing, so I add
some simple logic to core.py so that if .cif or .structure get a structure of
that type we simply return it.