Improve source detection
It is useful to sometimes simplify the species assignments. For example, if
there is more than one assigned species we should just assign it to synthetic. We
also should be able to filter by id in the query for source. I've added this
functionality as well as made some tests unskipped.