Add an first pass at a Motif Release loader
This loader will determine the information about the release to generate and
then runs the matlab code to cluster motifs. This loader will accept a list of
pdbs and then filter them to only the representative ones for use in clustering.
From here it then selects the valid loops from the best chains and models. These
loops are then clustered in matlab. The clustering creates some directories
which will be used later for import of data.

The clustering module is added as well. This is not yet tested but is mostly a
copy of the previous one, just organized a bit differently. Future commits will
add some smoke tests (at least) for that module as well.