Add an initial loader for chain to species mapping
It is really useful to store the assigned species for each chain. The chain info
table contains the mapping from chain to organism which may not be at the
species level. It may also include multiple entries for one chain. This table
will contain the simplified mapping we want to use. So it will be at the species
level and if given multiple things it will be mapped to synthetic construct. The
loader isn't tested super well yet, but is a start. It will be improved as we
use it more.