Add a base importer for motif data
Much motif data is basically a csv file that needs to be parsed. This can often
be abstracted into a single class. I think. This commit adds such a single class
and a few very basic tests for it. I've also added the required test data for
the motif data importers.