Add a basic strict mode for rna chains
Not all chains labeled as rna are rna chains. This is bad. I've added a not so
good strict mode which will filter out things which are not RNA. It's not ideal
as we will filter out things we do not want to get rid of as well. Eventually,
I'll improve this to only the requested chains.