Do not use transform to iterate
Anton suggested removing transform entirely, but I think it is actually useful
in some cases. However, it should not be used to generate values to iterate over
for processing as this abstraction is too much. Instead it is just better to
work on a single pdb at a time. This is a pretty major change to how all stages
run.