Start work on rewriting correspondence stuff
Correspondences will now be used to build the NR set instead of depending upon
them. This means we need to change the logic of what is selected for storage.
The results of correspondence must also include new information, such as how
good the alignment was. This is key to generating the nr sets so we will just
store it right away.  In addition, we have changed the design of experimental
sequences leading to changes in how we query and generate the data for input to
correspondence classes.