Add a flag to skip marking
In some cases we do not actually process pdb ids, instead we are processing
other things and so we cannot use the mark_processed method. Thus to support
that I add a flag to allow the skipping of marking. I would like to change
things so we still mark in those cases, but that requires more complex logic.