Allow correspondences against the sequence itself
This is actually really useful in that we can create a view which easily gets
all unit to unit correspondences. The downside to this approach is that we use a
cross join which is possibly very expensive. Since we have few rna sequences and
we do add a where condition it shouldn't be too bad. Hopefully. If someone
decides to extend this stuff in the future then this will need to be considered.