Add a wrapper function to handle unicode issues
Matlab must takes strings and not unicode strings. This is annoying so I've
added a wrapper around calls to mlab that returns a function that will convert
all of it's positional arguments to strings. This will work for our use case but
may not be a perfect general solution as we pass along keyword arguments
unchanged.