Also store correlation data in reversed order
It is useful for getting the units in both orders to store the correlation data
in both directions. I can't seem to write a view which will provide the data in
both directions without this. This does increase the amount of data we store and
alters the relationship between the info table and position table, as now the
exp_seq_id1 may not the be the same as what is implied by exp_seq_position_id1
but it makes everything easier to use. We do not store the reverse if they
reverse would be the same as is currently stored.