Update summary to deal with duplicated entries
We now write the data in both directions to the positions table. Thus to compute
the summary data correctly, we need to add a group by to the query so we only
get one row for each position in the alignment. It doesn't matter which one so I
don't try to set that.