Compute interaction to interaction correspondence
This is a working loader that can store interaction to interaction
correspondences for all structures. It stores tons of data right now and may be
too large for use in production, but it does work. It is possible to reduce the
amount of space used by creating a table of exp_seq_pairs, I think and then
storing references to that I think. That is something to explore.

One important thing to note that is that this uses generators extensively
because of the amount of data. Without them, it puts 2 gigs of data into memory
for each ecoli ssu. That will not scale well to more and larger structures.

I also added some more helper methods in the correspondence helper to make this
possible. The pdb_mapping method is nearly the same as the mapping method
because I can't dynamically generate the sql easily since sqlalchemy doesn't
reflect all views for some reason.