Add a topological sorting utility
This is copied from online, specifically:

https://bitbucket.org/ericvsmith/toposort/src/f45b9c78e8cb95db8f1bda9fd110d3d465c76d76/toposort.py?at=default#

I've added this as part of the source and modified it as it did not work with
python 2.6 because it used dictionary comprehensions. This has been fixed so I
can use it with 2.6. I've retained the copy-right notice so this should be above
board.