Add a dependency attribute to all Stages
Stages will now have dependencies. We add an empty set, ie nothing, as the
default one. This can be used in the dispatcher to make sure all things are run
as needed for the pipeline. A common source of errors is to try to run the
pipeline or part of it and realize that not everything got run or that things
failed because something should have been run first. The goal is to automate
this so that we don't need to run several things manually.