Add all dependency information
This now lets me use the depends data when running stages hopefully. I've mostly
mapped out all the dependencies need to make sure each stage runs correctly,
except any motif stuff and correspondence stuff as I'm rewriting that in another
branch. When these merge I'll deal with that.

I've also added a stage for creating matlab files. This is meant to make it
easier to track what is using matlab and what isn't. All matlab using things
will depend upon it to make it clear what is matlab and what isn't. We can't use
cifatom stuff as we are likely to provide that for downloads later on as well as
the current matlab usage.