Correct dependency definition and implement
I always forget that sets take an iterable not just the items, which is
annoying. Also, I've added a basic implementation and tests for this
functionality. Right now it defaults to off, but this may be changed in the
future. I've also changed MultiStageLoader to just be a simple stage that only
has dependencies on other stages.