Use improved query
The query now selects the pairs of experimental sequences in a more intelligent
way. We only pick things that could align well. In addition, we don't align
things that could be an SSU to things that could be an LSU. This cuts down on
the number of possible alignments quite a bit, making it feasible to do and
store all of them.