Detect if we have data when using a generator
We should be able to detect if we saved when given a generator. In that case we
can't simply check if the returned object is true as a generator is true even if
it yields nothing. To deal with this we set a flag to true if we have saved
anything when iterating. If we have not saved, then we do what we would
otherwise do for missing data, warn or throw an exception as configured.