Rework much of experimental sequence logic
There were some issues before where not all experimental sequence mappings were
created. In addition, it was not enforced by the database that all entries in
the experimental sequence table would be unique. I've fixed these problems as
well as changed how some stages work to make the queries simpler and clearer.
For example, we now use a unique constraint on the md5 of each sequence to make
sure they are unique. A unique constraint can't be used on the sequences
themselves because of size. Positions are mapped by transforming to all known
experimental sequences and then processing those. We correct detection of
unmapped sequences as before it would not occur if any sequence was mapped, now
all must be mapped.