Correct issue species comparison
The logic behind making sure the species do not conflict was flawed. Notably, it
did not deal with the species_id being NULL. This alters things to allow for
this. As a result we do more, though not many more, alignments. These alignments
seem to happen often for small things or fragments.