Refactor and add autonomous chains to the database
This autonomous chains are very useful to store for later stages, such as the
discrepancy and nr set calculations. Right now the method of storing them is
terrible as it does things twice. I can probably fix this by filling both tables
at once.

I also moved around a whole bunch of loaders and utility stuff in order to make
the setup more reasonable for this new design.