Get discrepancy computation to use autonomous
The discrepancies are now computed using the autonomous information. This is
pretty useful and important step in doing the nr set work. This appears to work
but is maybe too lose with allowing failures. However, there are lots of reasons
why the discrepancy can't be calculated, missing base centers and a bad
alignment are the two main ones. It would be nice to indicate this information
more clearly, however.