Add a class to create names for nr groups
This can probably be generalized for naming motif as well but we'll deal with
that when it happens. Right now this will name the nr classes as described in
anton's paper. The full thing is not yet tested but the parts that have been do
work. I've also begun outlining the tests that must occur.