Remove old nr stuff
I'm removing this because the old functionality is being split between views in
the database and python. This will speed up the pipeline as we will no longer
have to compute all differences ahead of time. In addition, the old stuff is
really inefficient and can be replaced with simpler set comparisons. I may be
copying old bits over into the new setup as needed, but for that I will just get
it from another branch.