Improve and test the naming system
The naming should now match the old method pretty closely. I've tested this by
comparing it to a release that has a large change. I may try to increase the
number of tests so it compares to many releases. I've also changed things so the
known handles are passed in, this way it is easier to reuse later as well as
test.