Correct dependencies of MutliStageLoader
The idea is that these classes should have the same dependencies as all stages
that this loader runs. This is determined by creating a set that merges all
stages this runs. In order to do it we have to add a simple read-only
classproperty decorator. I've also added a few simple tests to make sure of
this.