Correct a failing test
This fixes on test that was failing because we changed the meaning of a missing
annotation. Before it would return synthetic, now it gives None. One test is
still failing because I have yet to fix this issue. It does not yet know what to
do if the taxon id and name disagree. Thus I leave the failing test there.