Naming classes with new system
The idea for the new naming system is that we only have one handle per 'type'.
That is all resolution cutoffs of the same group have the same handle. This
makes it easier to look across cutoffs and know what an nr set contains. The way
this works is to do the naming based upon the 'all' cutoff and then change the
resolution part of each name.