Load resolution and adding sorting
We need to know resolution used in the groups for later usage. This is now
loaded along with the other autonomous group information. Also I've added some
tests for something that can rank the chains in a group. This is useful since we
have to have them ordered eventually as well.