Update to allow nr set importing
This appears to correctly load and import nr sets, with the possible exception
of loading parent data. I'm unsure on this because my local copy is messed up
right now. In addition, I've likely broken many tests and this will need to be
fixed as well. However, this does import nr clusterings now which is a big step
forward.