Change skipping logic
This would skip a whole structure if one type of loop is not found. This is
incorrect behavior and will lead to skipped loops in smaller structures. This
changes the skipping so that we only skip if all structures are missing all loop
types. Otherwise it will just warn for each missing loop type.