Try to improve performance of simple loaders
Simple loaders can take a while to check if there are any entries for a given
query. This is because it counts all matches which can be slow. We don't need a
full count however, just a 1 or 0. Thus we add a limit(1) clause to the query.
This should cut down on overall run time of these queries and improve pipeline
speed.