First attempt at renaming all columns
Many columns have changed names this attempts to make all the needed changes.
This is untested but does seem to touch all the major renames. It is possible
there are some missing changes. For example, some things use the results of
queries in clever ways and they may not have been renamed yet.