Simplify MultiStageLoaders
These loaders now exist only as a container for other loaders. I could also
possibly simplify things to only use a container actually. But as it is now they
just collect a series of other loaders and these are run. The MultiStageLoader
is not actually run in fact it is excluded from running at all. The stages
within it are run in the order defined by their dependencies. This simplifies
logic of figuring out what to run and in what order as it is all in one place
now, the dispatcher.