Add some tests for correspondence.info
Before it was hard to understand and untested. This way more logic is in python
and it is much easier to test. I have added some tests to make sure it respects
length and species assignments. The code could use a little clean up, hard coded
values and all but overall it is an improvement I think.