Deal with a possible error case
In some cases we cannot create the directory to store loops. This is a problem
that was not dealt with before. The cause is if the parent directory is not
present then it will not be created. Python's os.mkdir is not recursive and
os.mkdirs is not present in the version we use.