Update tests to deal with new data
We use different data sets now so we need to reflect this. We also simplify the
tests a bit in that we no longer check the chain_id as it may change depending
on the order of operations. The species id are stable however. Also, that is
more important for the tests that we are doing now.