Try to rework saving
Saving things to file is pretty messy right now. This is an attempt to fix that
by adding the Saver abstraction. I'm not yet sure if it is worth the extra
complexity but it may be helpful. We create an abstraction where saving is done
by calling some object. It handles the messy details of only saving so many
things at once (a limit in the DB) and making sure the file exists and is non
empty afterwards (a problem with writing to files). It can deal with all the
things I used before, objects, lists and generators, as well as dictionaries
instead of database objects. This may be useful when dealing with database stuff
(more informative logging statements) and is required for saving to files.

This also comes with a large regorinzation of core.py. It should be split up
anyway as it is pretty large (~700 lines) and was getting complicated anyway. I
haven't fixed all the issues caused by the split yet. This was just to make sure
the basic saving and importing functionality within core has been maintained.