Some corrects for new setup
The queries were not correct in that they produced too many chain groups. We
also have an issue where we do not get all the correct numbers of chains and
such. I think the tests are failing because of changes to files, or the lack of
imported interactions from the bootstrap script. I may try to import some
manually and see if that helps.