Update to new naming system
I move back to using the older style where grouping and mapping is passed to
chains. I think this is much more testable and cleaner. We also correct the
tests so we do not have the old test for .ids. I've commented them out because I
think they now belong elsewhere but I'm not sure yet. Also, we have corrected
the tests that can be run so they use the new data structures. We now use a
simpler setup of just a list of dictionaries instead of all sorts of mapping and
the like.