Add some more inferred locations
There are some instances, dealing with matlab for example, where knowing the
base directory will be helpful. This adds those locations to the defaults in
config. While this means they can be overridden leading to issues I feel that is
ok as it is easy to just not use that value.