Correct some issues in loading obsolete data
We want to only store a string in replaced_by if there was a replacing pdb. If
there isn't we should use NULL or None in python. Also, I simplify much of the
logic of the loader. No need to build a new list of database objects, we can
just use the newer saver features. Also, the stage may not produce no data so no
need for the final error logging statement.