Refactoring of correspondence info
I like this style more. The logic for selecting things is in python, not the
database and is easier for me to test. It does not use the database nearly as
much. Also it will allow me to see how many pairs will be created very easily.
It also will only create pairs between the given pdbs unlike the previous
version. I'm unsure about overall performance but it shouldn't be too bad.