Simplify loop extractor and fix tests
We no longer need to worry about old style ids so that chunk can be removed.
Also we can now test all the loop extractor so we should. For some reason the
tests are super slow. I'm not clear why but I don't know what to do about it
yet. We also correct some issues that arose during naming new loops. They would
not be sequential by loop type. Instead they would be sequential for the total
counts of loops.