Remove most code in release utilities
We don't need it to do so much code in this. It should do less. In particular
the code to get previous is always unneeded and overly complicated, thus it was
removed. In addition, we remove the code to get the current one. While this ends
up with some cut and paste code in places I think it makes things overall much
simpler.