Try to use new chains
When we migrated the database from old style to the new style we created some
IFE entries. These have NULL values for many columns, notably length, as that is
not computed as part of their workflow. This may change in the future but for
now there is no other information for them. By requiring a length they should be
prevented them from being used in new style NR sets. We do not want to use them
as those chains may not be IFE's as we now define them.