Begin rewrite of loop quality
The loop quality logic has to be reworked. It relied upon reading PDB files. As
these do not exist for all files anymore this cannot be used. We are going to
rewrite it in python to read cif files. It shouldn't be too bad, I hope. Most of
the data is easily computed or already stored in our database.